AK: Add Coroutine support and a simple Promise + Task type
Add equivalent declarations to <coroutine> that works with both
clang++ and g++ coroutine implementations. Unfortunately, the
coroutine_traits and coroutine_handle types are required to be in ::std.

Also add a simple Task and Promise type for simple coroutines that
simply return a value synchronously in an infallible way, and a test for
them. The Task type is inspired by cppcoro and a few other examples from
around the net.