LibGUI: Peel off a base class for ConnectionToWindowServer
This breaks ... many rules of serenity code, but will allow for a crafty
developer to implement a custom ConnectionToWindowServer that intends
to talk a different IPC language to a different window server without
changing any LibGUI-based applications.

A reasonable implementation of ConnectionToWindowServer::the() will
require reaching out to the GUI::Application to figure out which window
server backend we're supposed to instantiate.